requirements.txt
python-dotenv==0.19.0
requests==2.26.0
xmltodict==0.12.0
